Import common styles in filter
