+++ src/component/features/Filter/filterByContinent.scss
@@ -1,0 +1,10 @@
+@import '../../common/scss/variable.scss';
+@import '../../common/scss/mixins.scss';
+@import '../../common/scss/typography.scss';
+
+.continents-container {
+  width: 100%;
+  height: 90vh;
+  @include flex;
+  @include flex-center(row);
+}
